Simple Math game project has been updated here.
Mainly it's a game for kids. Two number pop up in the windon and you have to choose the correct answer from the options. If you guess the correct answer, then you will be rewarded with +1 score. The more you can answer correctly, the more your score will be within one minute.